requirements.txt
# requirements.txt
tensorflow
tensorflow-datasets
pyyaml
